graphql: Introduce execution::Query and move complexity calculation there
