+++ graphql/src/execution/mod.rs
@@ -4,5 +4,8 @@ mod execution;
 /// Common trait for field resolvers used in the execution.
 mod resolver;
 
+mod query;
+
 pub use self::execution::*;
+pub use self::query::Query;
 pub use self::resolver::{ObjectOrInterface, Resolver};
