+++ graphql/tests/query.rs
@@ -807,7 +807,7 @@ async fn query_complexity_subscriptions() {
 
     // This query is exactly at the maximum complexity.
     // FIXME: Not collecting the stream because that will hang the test.
-    let _ignore_stream = execute_subscription(&Subscription { query }, options).unwrap();
+    let _ignore_stream = execute_subscription(Subscription { query }, options).unwrap();
 
     let query = Query::new(
         Arc::new(api_test_schema()),
@@ -843,7 +843,7 @@ async fn query_complexity_subscriptions() {
     };
 
     // The extra introspection causes the complexity to go over.
-    let result = execute_subscription(&Subscription { query }, options);
+    let result = execute_subscription(Subscription { query }, options);
     match result {
         Err(SubscriptionError::GraphQLError(e)) => match e[0] {
             QueryExecutionError::TooComplex(1_010_200, _) => (), // Expected
@@ -1179,7 +1179,7 @@ async fn subscription_gets_result_even_without_events() {
 
     // Execute the subscription and expect at least one result to be
     // available in the result stream
-    let stream = execute_subscription(&Subscription { query }, options).unwrap();
+    let stream = execute_subscription(Subscription { query }, options).unwrap();
     let results: Vec<_> = stream
         .take(1)
         .collect()
